Dropped `BasicEnergyStorage` and `NetworkComponent`. Refactored energy API with long-based storage and updated component implementations. Initial implementation of A* for energy pathfinding (still WIP). Also additional Documentation

+++ src/main/java/de/kingsword/energy_api/capability/EnergyConsumer.java
@@ -1,5 +1,6 @@
 package de.kingsword.energy_api.capability;
 
 public interface EnergyConsumer extends EnergyCapability {
-    int consumeEnergy(int amount);
+    // delete energy from network
+    long consumeEnergy(int amount);
 }
