+++ src/main/java/de/kingsword/energy_api/capability/EnergyProducer.java
@@ -1,5 +1,6 @@
 package de.kingsword.energy_api.capability;
 
 public interface EnergyProducer extends EnergyCapability {
-    int produceEnergy(int amount);
+    // add energy to network
+    long produceEnergy(int amount);
 }
